+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInMetricsListener.cs
@@ -25,11 +25,13 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation;
 internal sealed class HttpInMetricsListener : IDisposable
 {
     private readonly Histogram<double> httpServerDuration;
+    private readonly AspNetMetricsInstrumentationOptions options;
 
-    public HttpInMetricsListener(Meter meter)
+    public HttpInMetricsListener(Meter meter, AspNetMetricsInstrumentationOptions options)
     {
         this.httpServerDuration = meter.CreateHistogram<double>("http.server.duration", "ms", "Measures the duration of inbound HTTP requests.");
         TelemetryHttpModule.Options.OnRequestStoppedCallback += this.OnStopActivity;
+        this.options = options;
     }
 
     public void Dispose()
@@ -47,6 +49,18 @@ internal sealed class HttpInMetricsListener : IDisposable
             { SemanticConventions.AttributeHttpScheme, context.Request.Url.Scheme },
             { SemanticConventions.AttributeHttpStatusCode, context.Response.StatusCode },
         };
+
+        if (this.options.Enrich is not null)
+        {
+            try
+            {
+                this.options.Enrich(context, ref tags);
+            }
+            catch (Exception ex)
+            {
+                AspNetInstrumentationEventSource.Log.EnrichmentException(nameof(HttpInMetricsListener), ex);
+            }
+        }
 
         this.httpServerDuration.Record(activity.Duration.TotalMilliseconds, tags);
     }