+++ src/OpenTelemetry.Instrumentation.AspNet/MeterProviderBuilderExtensions.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+using System;
 using OpenTelemetry.Instrumentation.AspNet;
 using OpenTelemetry.Internal;
 
@@ -29,12 +30,25 @@ public static class MeterProviderBuilderExtensions
     /// </summary>
     /// <param name="builder"><see cref="MeterProviderBuilder"/> being configured.</param>
     /// <returns>The instance of <see cref="MeterProviderBuilder"/> to chain the calls.</returns>
+    public static MeterProviderBuilder AddAspNetInstrumentation(this MeterProviderBuilder builder) =>
+        AddAspNetInstrumentation(builder, configure: null);
+
+    /// <summary>
+    /// Enables the incoming requests automatic data collection for ASP.NET.
+    /// </summary>
+    /// <param name="builder"><see cref="MeterProviderBuilder"/> being configured.</param>
+    /// <param name="configure">Callback action for configuring <see cref="AspNetMetricsInstrumentationOptions"/>.</param>
+    /// <returns>The instance of <see cref="MeterProviderBuilder"/> to chain the calls.</returns>
     public static MeterProviderBuilder AddAspNetInstrumentation(
-        this MeterProviderBuilder builder)
+        this MeterProviderBuilder builder,
+        Action<AspNetMetricsInstrumentationOptions>? configure)
     {
         Guard.ThrowIfNull(builder);
 
+        var options = new AspNetMetricsInstrumentationOptions();
+        configure?.Invoke(options);
+
         builder.AddMeter(AspNetMetrics.InstrumentationName);
-        return builder.AddInstrumentation(() => new AspNetMetrics());
+        return builder.AddInstrumentation(() => new AspNetMetrics(options));
     }
 }