+++ test/OpenTelemetry.Instrumentation.AspNet.Tests/HttpInMetricsListenerTests.cs
@@ -14,8 +14,9 @@
 // limitations under the License.
 // </copyright>
 
+using System;
 using System.Collections.Generic;
-using System.IO;
+using System.Diagnostics;
 using System.Threading;
 using System.Web;
 using OpenTelemetry.Context.Propagation;
@@ -27,14 +28,28 @@ namespace OpenTelemetry.Instrumentation.AspNet.Tests;
 
 public class HttpInMetricsListenerTests
 {
-    [Fact]
-    public void HttpDurationMetricIsEmitted()
+    [Theory]
+    [InlineData("http://localhost/", 0, null, null, "http")]
+    [InlineData("https://localhost/", 0, null, null, "https")]
+    [InlineData("http://localhost/api/value", 0, null, null, "http")]
+    [InlineData("http://localhost/api/value", 1, "{controller}/{action}", null, "http")]
+    [InlineData("http://localhost/api/value", 2, "{controller}/{action}", null, "http")]
+    [InlineData("http://localhost/api/value", 3, "{controller}/{action}", null, "http")]
+    [InlineData("http://localhost/api/value", 4, "{controller}/{action}", null, "http")]
+    [InlineData("http://localhost:8080/api/value", 0, null, null, "http")]
+    [InlineData("http://localhost:8080/api/value", 1, "{controller}/{action}", null, "http")]
+    [InlineData("http://localhost:8080/api/value", 3, "{controller}/{action}", "enrich", "http")]
+    [InlineData("http://localhost:8080/api/value", 3, "{controller}/{action}", "throw", "http")]
+    [InlineData("http://localhost:8080/api/value", 3, "{controller}/{action}", null, "http")]
+    public void AspNetMetricTagsAreCollectedSuccessfully(
+        string url,
+        int routeType,
+        string routeTemplate,
+        string enrichMode,
+        string expectedScheme)
     {
-        string url = "http://localhost/api/value";
         double duration = 0;
-        HttpContext.Current = new HttpContext(
-            new HttpRequest(string.Empty, url, string.Empty),
-            new HttpResponse(new StringWriter()));
+        HttpContext.Current = RouteTestHelper.BuildHttpContext(url, routeType, routeTemplate);
 
         // This is to enable activity creation
         // as it is created using ActivitySource inside TelemetryHttpModule
@@ -52,7 +67,21 @@ public class HttpInMetricsListenerTests
 
         var exportedItems = new List<Metric>();
         using var meterProvider = Sdk.CreateMeterProviderBuilder()
-            .AddAspNetInstrumentation()
+            .AddAspNetInstrumentation(options =>
+            {
+                options.Enrich += (HttpContext context, ref TagList tags) =>
+                {
+                    if (enrichMode == "throw")
+                    {
+                        throw new Exception("Enrich exception");
+                    }
+
+                    if (enrichMode == "enrich")
+                    {
+                        tags.Add("enriched", "true");
+                    }
+                };
+            })
             .AddInMemoryExporter(exportedItems)
             .Build();
 
@@ -61,6 +90,8 @@ public class HttpInMetricsListenerTests
         ActivityHelper.StopAspNetActivity(Propagators.DefaultTextMapPropagator, activity, HttpContext.Current, TelemetryHttpModule.Options.OnRequestStoppedCallback);
 
         meterProvider.ForceFlush();
+
+        Assert.Single(exportedItems);
 
         var metricPoints = new List<MetricPoint>();
         foreach (var p in exportedItems[0].GetMetricPoints())
@@ -81,34 +112,43 @@ public class HttpInMetricsListenerTests
         Assert.Equal(duration, sum);
         Assert.True(duration > 0, "Metric duration should be set.");
 
-        Assert.Equal(3, metricPoints[0].Tags.Count);
-        string? httpMethod = null;
-        int httpStatusCode = 0;
-        string? httpScheme = null;
-
-        foreach (var tag in metricPoints[0].Tags)
+        var expectedTagCount = 3;
+        if (enrichMode == "enrich")
         {
-            if (tag.Key == SemanticConventions.AttributeHttpMethod)
-            {
-                httpMethod = (string)tag.Value;
-                continue;
-            }
-
-            if (tag.Key == SemanticConventions.AttributeHttpStatusCode)
-            {
-                httpStatusCode = (int)tag.Value;
-                continue;
-            }
-
-            if (tag.Key == SemanticConventions.AttributeHttpScheme)
-            {
-                httpScheme = (string)tag.Value;
-                continue;
-            }
+            expectedTagCount++;
         }
 
-        Assert.Equal("GET", httpMethod);
-        Assert.Equal(200, httpStatusCode);
-        Assert.Equal("http", httpScheme);
+        Assert.Equal(expectedTagCount, metricPoints[0].Tags.Count);
+        Dictionary<string, object?> tags = new(metricPoint.Tags.Count);
+        foreach (var tag in metricPoint.Tags)
+        {
+            tags.Add(tag.Key, tag.Value);
+        }
+
+        if (enrichMode == "enrich")
+        {
+            ExpectTag("true", "enriched");
+        }
+
+        ExpectTag("GET", SemanticConventions.AttributeHttpMethod);
+        ExpectTag(200, SemanticConventions.AttributeHttpStatusCode);
+        ExpectTag(expectedScheme, SemanticConventions.AttributeHttpScheme);
+
+        void ExpectTag<T>(T? expected, string tagName)
+        {
+            if (expected is null)
+            {
+                Assert.DoesNotContain(tagName, tags.Keys);
+                return;
+            }
+
+            if (tags.TryGetValue(tagName, out var value))
+            {
+                Assert.Equal(expected, (T?)value);
+                return;
+            }
+
+            Assert.Fail($"Expected tag with key {tagName} not found.");
+        }
     }
 }